+++ src/js/main.js
@@ -2,7 +2,6 @@
  * Module Description
  *
  * @author Author Name
- * @date 2013-01-01
  */
 
 require([