+++ src/modules/module/module.js
@@ -2,7 +2,6 @@
  * An example module
  *
  * @author Author name
- * @date 2013-08-18
  */
 
 define([