+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ### HEAD
+* Remove unused `@date` definitions
 * Update Apache Server Configs to v2.14.0
 * Remove node dependency `xtend` in favor of `node-deep-extend` (#132)
 * Add HTML to modules folder and parse module html (#129)